requirements.txt
bottle
jinja2
ZODB
ZEO
M2Crypto
zope.component
zope.interface
zodbpickle
gunicorn
passlib
